api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
mediapipe>=0.10.14
opencv-python-headless>=4.9.0
anthropic>=0.28.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic-settings>=2.0.0
